+++ testing/Server/DataKeepTest.server.lua
@@ -51,101 +51,113 @@ end
 --> Public Functions
 
 Players.PlayerAdded:Connect(function(player)
-	keepStore:LoadKeep("Player_" .. player.UserId):andThen(function(keep)
-		Keeps[player] = keep
+	-- keepStore:LoadKeep("Player_" .. player.UserId):andThen(function(keep)
+	-- 	Keeps[player] = keep
 
-		-- keep:onDataChange("Inventory.Sword", function(newValue)
-		-- 	print("Sword changed to", newValue)
-		-- end)
+	-- 	-- keep:onDataChange("Inventory.Sword", function(newValue)
+	-- 	-- 	print("Sword changed to", newValue)
+	-- 	-- end)
 
-		-- keep:Mutate("Inventory.Sword", function()
-		-- 	return "Iron Sword 2"
-		-- end)
+	-- 	-- keep:Mutate("Inventory.Sword", function()
+	-- 	-- 	return "Iron Sword 2"
+	-- 	-- end)
 
-		keep:CoolFunction()
-		--keep.Data.Test = nil
+	-- 	keep:CoolFunction()
+	-- 	--keep.Data.Test = nil
+
+	-- 	print(keep.Data)
+
+	-- 	keep:Reconcile()
+
+	-- 	print(keep.Data)
+
+	-- 	keep.OnGlobalUpdate:Connect(function(_, id)
+	-- 		print(player.UserId .. " received global update", id)
+
+	-- 		keep:ClearLockedUpdate(id)
+	-- 	end)
+
+	-- 	print("Lets try loading AGAIN")
+
+	-- 	keepStore:LoadKeep("Player_" .. player.UserId):andThen(function(keep2)
+	-- 		print("Loaded cached keep again")
+	-- 		print(keep2, keep)
+	-- 	end)
+
+	-- 	keep:Save()
+
+	-- 	-- keep.Data.Test = "Hello World! 2"
+	-- 	-- keep:Save()
+
+	-- 	-- local versions = keep:GetVersions()
+
+	-- 	-- versions:andThen(function(iterator)
+	-- 	-- 	print(keep.Data)
+	-- 	-- 	local versionToRoll = iterator.Current()
+
+	-- 	-- 	keep:SetVersion(versionToRoll.Version)
+
+	-- 	-- 	print(keep.Data)
+	-- 	-- end)
+
+	-- 	-- print("UserIds: ")
+	-- 	-- keep:AddUserId(player.UserId)
+
+	-- 	-- print(keep.UserIds)
+
+	-- 	-- keep:AddUserId(player.UserId)
+
+	-- 	-- print(keep.UserIds)
+
+	-- 	warn(keep.Data)
+
+	-- 	print(keepStore._store:GetAsync(keep._key))
+
+	-- 	-- keep:Release()
+	-- end)
+
+	-- for i = 1, 2 do
+	-- 	local message = if i == 1 then "Hello" else "Goodbye"
+
+	-- 	keepStore:PostGlobalUpdate("Player_" .. player.UserId, function(globalUpdates)
+	-- 		print(globalUpdates:GetActiveUpdates())
+	-- 		for _, globalUpdate in ipairs(globalUpdates:GetActiveUpdates()) do
+	-- 			print("Changing global update", globalUpdate.ID)
+
+	-- 			globalUpdates:ChangeActiveUpdate(globalUpdate.ID, {
+	-- 				Message = globalUpdate.Data.Message .. message,
+	-- 			})
+
+	-- 			print(globalUpdates:GetActiveUpdates())
+
+	-- 			return
+	-- 		end
+
+	-- 		print("Posted global update")
+
+	-- 		globalUpdates:AddGlobalUpdate({
+	-- 			Message = message,
+	-- 		})
+	-- 	end)
+	-- end
+
+	print("viewing")
+	keepStore:ViewKeep("Player_" .. player.UserId):andThen(function(keep)
+		print("Viewed keep")
+		print(keep.Data)
+
+		print("Changing")
+
+		keep:Mutate("Inventory.Sword", function()
+			return "Iron Sword 2"
+		end)
 
 		print(keep.Data)
 
-		keep:Reconcile()
-
-		print(keep.Data)
-
-		keep.OnGlobalUpdate:Connect(function(_, id)
-			print(player.UserId .. " received global update", id)
-
-			keep:ClearLockedUpdate(id)
-		end)
-
-		print("Lets try loading AGAIN")
-
-		keepStore:LoadKeep("Player_" .. player.UserId):andThen(function(keep2)
-			print("Loaded cached keep again")
-			print(keep2, keep)
-		end)
-
-		keep:Save()
-
-		-- keep.Data.Test = "Hello World! 2"
-		-- keep:Save()
-
-		-- local versions = keep:GetVersions()
-
-		-- versions:andThen(function(iterator)
-		-- 	print(keep.Data)
-		-- 	local versionToRoll = iterator.Current()
-
-		-- 	keep:SetVersion(versionToRoll.Version)
-
-		-- 	print(keep.Data)
-		-- end)
-
-		-- print("UserIds: ")
-		-- keep:AddUserId(player.UserId)
-
-		-- print(keep.UserIds)
-
-		-- keep:AddUserId(player.UserId)
-
-		-- print(keep.UserIds)
-
-		warn(keep.Data)
+		keep:Overwrite()
 
 		print(keepStore._store:GetAsync(keep._key))
-
-		-- keep:Release()
 	end)
-
-	for i = 1, 2 do
-		local message = if i == 1 then "Hello" else "Goodbye"
-
-		keepStore:PostGlobalUpdate("Player_" .. player.UserId, function(globalUpdates)
-			print(globalUpdates:GetActiveUpdates())
-			for _, globalUpdate in ipairs(globalUpdates:GetActiveUpdates()) do
-				print("Changing global update", globalUpdate.ID)
-
-				globalUpdates:ChangeActiveUpdate(globalUpdate.ID, {
-					Message = globalUpdate.Data.Message .. message,
-				})
-
-				print(globalUpdates:GetActiveUpdates())
-
-				return
-			end
-
-			print("Posted global update")
-
-			globalUpdates:AddGlobalUpdate({
-				Message = message,
-			})
-		end)
-	end
-
-	-- print("viewing")
-	-- keepStore:ViewKeep("Player_" .. player.UserId):andThen(function(keep)
-	-- 	print("Viewed keep")
-	-- 	print(keep.Data)
-	-- end)
 end)
 
 Players.PlayerRemoving:Connect(function(player)