+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # DataKeep
 
+## [version 2.1.2](https://github.com/noahrepublic/DataKeep/releases/tag/v2.1.2): 12/17/2023
+
+## **API Breaking Change**
+
+### Fixed
+- ViewKeep :Save not working, replaced with :Overwrite()
+
+### Added
+- ViewKeep :Overwrite()
+
 ## [version 2.1.1](https://github.com/noahrepublic/DataKeep/releases/tag/v2.1.1): 12/05/2023
 
 ### Fixes